adjustments to employee data and repositories. And migration of population from the tables employees and addresses

+++ Dell.Lead.WeApi/Business/IEmployeeBusiness.cs
@@ -5,11 +5,11 @@ namespace Dell.Lead.WeApi.Business
 {
     public interface IEmployeeBusiness
     {
-        EmployeeVO FindByCpf(int cpf);
+        EmployeeVO FindByCpf(long cpf);
         List<EmployeeVO> FindAll();
         EmployeeVO Create(EmployeeVO employee);
         EmployeeVO Update(EmployeeVO employee);
-        void Delete(int cpf);
+        void Delete(long cpf);
 
     }
 }
