+++ Dell.Lead.WeApi/Data/VO/EmployeeVO.cs
@@ -6,9 +6,9 @@ namespace Dell.Lead.WeApi.Data.VO
     {
         public long Id { get; set; }
         public string NameFull { get; set; }
-        public int Cpf { get; set; }
+        public long Cpf { get; set; }
         public DateTime BirthDate { get; set; }
-        public int Phone { get; set; }
+        public long Phone { get; set; }
         public string Gender { get; set; }
         public AddressVO Address { get; set; }
     }
